% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaptiveIndex.function.R
\name{CaptiveIndex}
\alias{CaptiveIndex}
\title{CaptiveIndex Function}
\usage{
CaptiveIndex(x, Weight.1, Weight.2, Weight.3 = NULL, Weight.4 = NULL)
}
\arguments{
\item{Dataset}{name, Nearest neighbour weight, Close proximity weight, v3, v4}
}
\description{
This function calculates the sociality measure using the data created from the DataCreate Function.
}
\examples{
Results<-CaptiveIndex (NN.CP, 1.4, 53.2)
CaptiveIndex()
}
\keyword{CaptiveIndex,}
\keyword{association}
\keyword{index,}
\keyword{weightings}
