% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCreate.function.R
\name{DataCreate}
\alias{DataCreate}
\title{DataCreate function}
\usage{
DataCreate(x, Measures = NULL)
}
\arguments{
\item{Add}{in the dataframe matching the example using the appropraite codes}
}
\description{
This function allows you to create a dataset to be used in the captive association index
}
\examples{
NN.CP <- DataCreate (giraffe)
head (NN.CP)
write.csv(NN.CP, "Results.csv")
DataCreate()
}
\keyword{Captive}
\keyword{association}
\keyword{index}
