% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\name{Weights}
\alias{Weights}
\title{Weights function}
\usage{
Weights(x)
}
\arguments{
\item{x}{is the dataframe matching the example using the appropriate column headers (see details)}
}
\description{
This function calculates the weights for the proximity conditions
}
\examples{
Weights (giraffe)
}
\keyword{conditions}
\keyword{nearest}
\keyword{neighbour,}
\keyword{proximity}
