% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaptiveIndex.function.R
\name{CaptiveIndex}
\alias{CaptiveIndex}
\title{CaptiveIndex Function}
\usage{
CaptiveIndex(x, Weight.1, Weight.2)
}
\arguments{
\item{x}{is the output dataframe from the /code{DataCreate} function}

\item{Weight.1}{Input Nearest Neighbour Weight from /code{Weights} function}

\item{Weight.2}{Input Interaction Distance Weight from /code{Weights} function}
}
\description{
This function calculates the sociality measure using the data created from the DataCreate Function.
}
\examples{
data(giraffe)
W<-Weights (giraffe)
NN.ID <- DataCreate (giraffe)
head (NN.ID)
Results<-CaptiveIndex (NN.ID, W$NN, W$IDist)
head (Results)
}
\keyword{CaptiveIndex,}
\keyword{association}
\keyword{index,}
\keyword{weights}
