% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCreate.function.R
\name{DataCreate}
\alias{DataCreate}
\title{DataCreate function}
\usage{
DataCreate(x, Measures = NULL)
}
\arguments{
\item{x}{is the dataframe matching the example using the appropriate column headers (see details)}
}
\description{
This function allows you to create a dataset to be used in the captive association index
}
\examples{
NN.ID <- DataCreate (giraffe)
head (NN.ID)
}
\keyword{Association}
\keyword{Captive}
\keyword{Index,}
\keyword{index}
