% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\name{Weights}
\alias{Weights}
\title{Weights function}
\usage{
Weights(x)
}
\arguments{
\item{add}{in dataframe names}
}
\description{
This function calculates the weighting for nearest neighbour and proximity values
}
\examples{
Weights (giraffe)
Weights()
}
\keyword{and}
\keyword{nearest}
\keyword{neighbour}
\keyword{proximity}
\keyword{values}
\keyword{weights}
